% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagine-main.R
\docType{package}
\name{image-package}
\alias{image-package}
\alias{imagine}
\alias{imagine-package}
\title{Imaging Engine, Tools for application of image filters to data matrices}
\description{
Package for faster application of image filters to data matrices, using R and C++ algorithms.
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@gmail.com}
}
\keyword{image-filter}
\keyword{image-matrix,}

