% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagine-main.R
\name{medianFilter2D}
\alias{medianFilter2D}
\title{Make a 2D median filter calculation from numeric matrix}
\usage{
medianFilter2D(dataMatrix, radius, times = 1)
}
\arguments{
\item{dataMatrix}{A \code{numeric matrix} object used for apply filters.}

\item{radius}{Size of squared kernel to apply median.}

\item{times}{How many times do you want to apply the filter?}
}
\value{
It returns a \code{matrix} object with the same dimensions of \code{dataMatrix}.
}
\description{
This function takes a \code{matrix} object, and for each cell multiplies its neighborhood by
the squared matrix of dimension \eqn{radius*radius}. Finally, it returns for each cell the median of the
weighted sum.
}
\details{
This function uses the \code{engine2} C++ algorithm. More details are shown in vignette.
}
\examples{
# Generate example matrix
nRows <- 10
nCols <- 20

myMatrix <- matrix(runif(nRows*nCols, 0, 100), nrow = nRows, ncol = nCols)
radius <- 3

# Make convolution
myOutput <- medianFilter2D(myMatrix, radius)

# Plot results
image(myMatrix, zlim = c(0, 100))
image(myOutput, zlim = c(0, 100))
}

