% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagine-main.R
\name{agenbagFilters}
\alias{agenbagFilters}
\title{Performs gradient algorithms from Agenbag et al. (2003)}
\usage{
agenbagFilters(X, algorithm = c(1, 2))
}
\arguments{
\item{X}{A numeric \code{matrix} object used for apply filters.}

\item{algorithm}{\code{integer} Specifying the type of method that will be
used. See Details.}
}
\value{
\code{agenbagFilters} returns a \code{matrix} object with the same
dimensions of \code{X}.
}
\description{
This function performs two gradient calculation approaches for
SST, as outlined in the paper by Agenbag et al. (2003).
}
\details{
Section 2.2.4 of the paper by Agenbag et al. (2003) introduces the following
two methods:
\itemize{
 \item{Method 1: }{Based on the equation
 \deqn{SST_{grad}=\sqrt{(T_{i+1}-T_{i-1})^2 +(T_{j+1}-T_{j-1})^2}}}
 \item{Method 2: }{the standard deviation of SST in a 3 x 3 pixel area
 centered on position (i,j).}
}
}
\examples{
data(wbImage)

# Agenbag, gradient algorithm 1
agenbag1 <- agenbagFilters(X = wbImage, algorithm = 1)

# image(agenbag1)

# Agenbag, gradient algorithm 2
agenbag2 <- agenbagFilters(X = wbImage, algorithm = 2)

# image(agenbag2)
}
\references{
Agenbag, J.J., A.J. Richardson, H. Demarcq, P. Freon, S. Weeks,
and F.A. Shillington. "Estimating Environmental Preferences of South African
Pelagic Fish Species Using Catch Size- and Remote Sensing Data". Progress in
Oceanography 59, No 2-3 (October 2003): 275-300.
(\doi{https://doi.org/10.1016/j.pocean.2003.07.004}).
}
