% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagine-main.R
\docType{package}
\name{image-package}
\alias{image-package}
\alias{imagine-package}
\alias{imagine}
\title{IMAGing engINE, Tools for application of image filters to data matrices}
\description{
Provides fast application of image filters to data matrices, using R and C++ algorithms.
}
\details{
This package uses C++ algorithms called 'engines'. More details are shown in the vignette.
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@gmail.com}
}
\concept{image-filter}
\concept{image-matrix}
