% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagine-package.R
\docType{package}
\name{imagine-package}
\alias{imagine}
\alias{imagine-package}
\title{IMAGing engINE, Tools for application of image filters to data matrices}
\description{
Provides fast application of image filters to data matrices,
using R and C++ algorithms.
}
\details{
This package uses C++ algorithms called 'engines'. More details are
shown in the vignette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LuisLauM/imagine}
  \item Report bugs at \url{https://github.com/LuisLauM/imagine/issues}
}

}
\author{
\strong{Maintainer}: Wencheng Lau-Medrano \email{luis.laum@gmail.com}

}
\keyword{internal}
