% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagine-main.R
\docType{data}
\name{wbImage}
\alias{wbImage}
\title{Data matrix to be used as example image.}
\format{A \code{matrix} with dimnensions 1280x720.}
\usage{
wbImage
}
\description{
\code{matrix} object containig numeric data to plot a image. The photo
was taken by the author at 2016.
}
